Cheval Down
This makes it to where when LStick But. 5 is pressed, it lowers the
Cheval actuator, but it must be held in order to prevent the isFinished
method from running. In effect, this allows us to have a whileHeld
button command that also has an end() method. TL;DR: I’m a genius.