Control Update
This is a pretty massive overhaul of how we control the lifter and
shooter. The shooter has three positions (1 being top and 3 being
bottom), and are cycled through using RStick But.’s 6 and 4. The lifter
is now controlled by a toggle on LStick But. 3. If this is broken, look
in the controller methods in DriveWithJoysticks